stock: Remove closeEnough and guard against negative cost.
I'm not sure closeEnough was added to begin with. An error margin of 0.02 is rather significant. Very little difference in output now. However, it was resulting in a negative cost basis in the final gains output which was bad.

Maybe an earlier version of the algorithm couldn't handle missing purchases as well. Now we just use the fallback price for a conservative cost basis.